+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ***Version 1.0.0.2011* ** *(TBD)*
 - Fixed a cosmetic bug in the Timer Setup Screen, where the succeeding timer could show an incorrect timer, if that timer had been deleted.
+- Added accessibility stuff.
 
 ***Version 1.0.0.2010* ** *(December 8, 2018)*
 - Fixed an issue where there was inconsistent behavior between touching the pause/play button, and tapping in the screen for Podium Mode (the behavior should be the same).